requirements.txt
pandas==2.1.3
tabulate==0.9.0
requests==2.31.0
python-telegram-bot==20.6
selenium==4.15.2